Get Tabs working in the settings page
