+++ services/app-web/src/components/Tabs/Tabs.stories.tsx
@@ -15,50 +15,35 @@ export default {
     },
 }
 
-const TabPanelWrapper: React.FC<
-    React.PropsWithChildren<
-        React.PropsWithChildren<React.PropsWithChildren<unknown>>
-    >
-> = ({ children }) => {
-    return <div style={{ margin: '10px', height: '500px' }}>{children}</div>
-}
 export const Default = (): React.ReactElement => {
     return (
         <Tabs>
             <TabPanel id="Test-Pepperoni" tabName="Pepperoni">
-                <TabPanelWrapper>
-                    <h1>Pepperoni</h1>
-                    <p>
-                        Great Pizza! A tale of two pizzas. Pizza now and in the
-                        future.
-                    </p>
-                </TabPanelWrapper>
+                <h1>Pepperoni</h1>
+                <p>
+                    Great Pizza! A tale of two pizzas. Pizza now and in the
+                    future.
+                </p>
             </TabPanel>
             <TabPanel id="Test-Sausage" tabName="Sausage">
-                <TabPanelWrapper>
-                    <h1>Sausage</h1>
-                    <p>Great Pizza!</p>
-                </TabPanelWrapper>
+                <h1>Sausage</h1>
+                <p>Great Pizza!</p>
             </TabPanel>
             <TabPanel id="Test-Mushroom" tabName="Mushroom">
-                <TabPanelWrapper>
-                    <h1>Mushroom</h1>
-                    <p>Great Pizza!</p>
-                </TabPanelWrapper>
+                <h1>Mushroom</h1>
+                <p>Great Pizza!</p>
             </TabPanel>
             <TabPanel id="Test-Bacon" tabName="Bacon">
-                <TabPanelWrapper>
-                    <h1>Bacon</h1>
-                    <p>Great Pizza!</p>
-                    <p>
-                        Great Pizza! A tale of two pizzas. Pizza now and in the
-                        future.
-                    </p>
-                    <p>
-                        Great Pizza! A tale of two pizzas. Pizza now and in the
-                        future.
-                    </p>
-                </TabPanelWrapper>
+                <h1>Bacon</h1>
+                <p>Great Pizza!</p>
+                <p>
+                    Great Pizza! A tale of two pizzas. Pizza now and in the
+                    future.
+                </p>
+                <p>
+                    Great Pizza! A tale of two pizzas. Pizza now and in the
+                    future.
+                </p>
             </TabPanel>
         </Tabs>
     )
@@ -68,39 +53,31 @@ export const CustomDefaultTab = (): React.ReactElement => {
     return (
         <Tabs>
             <TabPanel id="Test-Pepperoni" tabName="Pepperoni">
-                <TabPanelWrapper>
-                    <h1>Pepperoni</h1>
-                    <p>
-                        Great Pizza! A tale of two pizzas. Pizza now and in the
-                        future.
-                    </p>
-                </TabPanelWrapper>
+                <h1>Pepperoni</h1>
+                <p>
+                    Great Pizza! A tale of two pizzas. Pizza now and in the
+                    future.
+                </p>
             </TabPanel>
             <TabPanel id="Test-Sausage" tabName="Sausage">
-                <TabPanelWrapper>
-                    <h1>Sausage</h1>
-                    <p>Great Pizza!</p>
-                </TabPanelWrapper>
+                <h1>Sausage</h1>
+                <p>Great Pizza!</p>
             </TabPanel>
             <TabPanel id="Test-Mushroom" tabName="Mushroom">
-                <TabPanelWrapper>
-                    <h1>Mushroom</h1>
-                    <p>Great Pizza!</p>
-                </TabPanelWrapper>
+                <h1>Mushroom</h1>
+                <p>Great Pizza!</p>
             </TabPanel>
             <TabPanel id="Test-Bacon" tabName="Bacon">
-                <TabPanelWrapper>
-                    <h1>Bacon</h1>
-                    <p>Great Pizza!</p>
-                    <p>
-                        Great Pizza! A tale of two pizzas. Pizza now and in the
-                        future.
-                    </p>
-                    <p>
-                        Great Pizza! A tale of two pizzas. Pizza now and in the
-                        future.
-                    </p>
-                </TabPanelWrapper>
+                <h1>Bacon</h1>
+                <p>Great Pizza!</p>
+                <p>
+                    Great Pizza! A tale of two pizzas. Pizza now and in the
+                    future.
+                </p>
+                <p>
+                    Great Pizza! A tale of two pizzas. Pizza now and in the
+                    future.
+                </p>
             </TabPanel>
         </Tabs>
     )
