+++ services/app-web/src/pages/Settings/CMSUsersTable/CMSUsersTable.tsx
@@ -1,0 +1,125 @@
+import { Table } from '@trussworks/react-uswds'
+import React, { useMemo } from 'react'
+import {
+    createColumnHelper,
+    flexRender,
+    getCoreRowModel,
+    useReactTable,
+} from '@tanstack/react-table'
+import {
+    CmsUser,
+    useIndexUsersQuery,
+    IndexUsersQuery,
+} from '../../../gen/gqlClient'
+
+import styles from '../Settings.module.scss'
+import { Loading } from '../../../components'
+import { SettingsErrorAlert } from '../SettingsErrorAlert'
+
+const columnHelper = createColumnHelper<CmsUser>()
+
+export const CMSUsersTable = (): React.ReactElement => {
+    const { loading, data, error } = useIndexUsersQuery({
+        fetchPolicy: 'network-only',
+    })
+
+    const showLoading = loading || !data
+
+    const columns = useMemo(
+        () => [
+            columnHelper.accessor('familyName', {
+                cell: (info) => info.getValue(),
+                header: () => 'Family Name',
+            }),
+            columnHelper.accessor('givenName', {
+                cell: (info) => info.getValue(),
+                header: () => 'Given Name',
+            }),
+            columnHelper.accessor('email', {
+                cell: (info) => info.getValue(),
+                header: () => 'Email',
+            }),
+        ],
+        []
+    )
+
+    // pick out the part of IndexUsersQuery that specifies Admin/CMS/StateUser
+    type UserTypesInIndexQuery = Pick<
+        IndexUsersQuery['indexUsers']['edges'][number],
+        'node'
+    >['node']
+
+    function isCmsUser(obj: UserTypesInIndexQuery): obj is CmsUser {
+        return obj.__typename === 'CMSUser'
+    }
+
+    const filterForCmsUsers = (
+        data: IndexUsersQuery | undefined
+    ): CmsUser[] => {
+        if (!data) {
+            return []
+        }
+        const cmsUsers = data.indexUsers.edges
+            .filter((edge) => isCmsUser(edge.node))
+            .map((edge) => edge.node as CmsUser)
+        return cmsUsers
+    }
+
+    const cmsUsers = filterForCmsUsers(data)
+
+    const table = useReactTable({
+        data: cmsUsers,
+        columns,
+        getCoreRowModel: getCoreRowModel(),
+    })
+
+    if (error) {
+        return <SettingsErrorAlert error={error} />
+    }
+
+    return (
+        <div className={styles.table}>
+            {showLoading ? (
+                <Loading />
+            ) : cmsUsers.length ? (
+                <Table bordered caption="CMS Users">
+                    <thead className={styles.header}>
+                        {table.getHeaderGroups().map((headerGroup) => (
+                            <tr key={headerGroup.id}>
+                                {headerGroup.headers.map((header) => (
+                                    <th key={header.id}>
+                                        {header.isPlaceholder
+                                            ? null
+                                            : flexRender(
+                                                  header.column.columnDef
+                                                      .header,
+                                                  header.getContext()
+                                              )}
+                                    </th>
+                                ))}
+                            </tr>
+                        ))}
+                    </thead>
+                    <tbody>
+                        {table.getRowModel().rows.map((row) => (
+                            <tr key={row.id}>
+                                {row.getVisibleCells().map((cell) => (
+                                    <td key={cell.id}>
+                                        {flexRender(
+                                            cell.column.columnDef.cell,
+                                            cell.getContext()
+                                        )}
+                                    </td>
+                                ))}
+                            </tr>
+                        ))}
+                    </tbody>
+                </Table>
+            ) : (
+                <div>
+                    <p>No CMS users to display</p>
+                </div>
+            )}
+        </div>
+    )
+}
