+++ services/app-web/src/pages/Settings/Settings.module.scss
@@ -1,10 +1,15 @@
+@import '../../styles/uswdsImports.scss';
+@import '../../styles/custom.scss';
+
 .table {
     width: 60%;
 }
-
+.tabs {
+    padding: 2em 0;
+}
 .pageContainer {
-    margin: units(2) 0;
-    align-items: start;
+    padding: 1em 10em;
+    width: 100%;
     background-color: #FFF;
 }
 .header {
